fix: cleanup mount ref
